mysite/requirements.txt
django==1.5.12
MySQL-python
django-xadmin
django-crispy-forms
